v1.002_addPlayerTeams
Team Sheets – Adding a player. Does it add to the team text file? We
couldn’t figure out where the team text files were stored to check this
ourselves. If it doesn’t already add it to the text file, is this
possible?
When adding a new player, upon clicking ‘Add to Squad’ can it be added
to the Squad list (where the text file loads) rather than the team sheet
box?
It’s possible to add a space (blank space) when clicking ‘adding to
squad’. Can a pop up box alert you upon doing this?